refactor: remove unused utils and redirect callers
