--- src/lib/utils/limitString.ts
@@ -1,5 +1,0 @@
-const limitString = (str: string, limit: number) => {
-    return str?.length > limit ? str.slice(0, limit) + '...' : str;
-};
-
-export default limitString;
